requirements.txt
# Core dependencies
hopsworks==4.2.*
# For Apache Arrow support (required for Feature Store)
pyarrow>=10.0.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# ML Models
xgboost>=1.7.0
lightgbm>=3.3.0

# API and HTTP requests
requests>=2.28.0
python-dotenv>=0.21.0

# Dashboard
flask>=2.3.0
flask-cors>=3.0.10

# Data processing
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.2

# Optional: for better data handling
joblib>=1.2.0

